+++ lib/screens/settings.dart
@@ -1,16 +1,18 @@
+import 'package:every_door/fields/combo.dart';
+import 'package:every_door/fields/helpers/combo_page.dart';
 import 'package:every_door/providers/changes.dart';
-import 'package:every_door/providers/editor_mode.dart';
 import 'package:every_door/providers/editor_settings.dart';
 import 'package:every_door/providers/osm_auth.dart';
 import 'package:every_door/providers/osm_data.dart';
-import 'package:every_door/screens/entrances.dart';
+import 'package:every_door/providers/presets.dart';
 import 'package:every_door/screens/settings/account.dart';
 import 'package:every_door/screens/settings/changes.dart';
 import 'package:every_door/screens/settings/imagery.dart';
+import 'package:every_door/widgets/mode_section.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dropdown_alert/alert_controller.dart';
 import 'package:flutter_dropdown_alert/model/data_alert.dart';
-import 'package:flutter_settings_ui/flutter_settings_ui.dart';
+import 'package:settings_ui/settings_ui.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -25,16 +27,9 @@ class SettingsPage extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final changes = ref.watch(changesProvider);
     final osmData = ref.watch(osmDataProvider);
-    final editorMode = ref.watch(editorModeProvider);
     final login = ref.watch(authProvider);
     final editorSettings = ref.watch(editorSettingsProvider);
     final loc = AppLocalizations.of(context)!;
-    final titlePadding = EdgeInsets.only(
-      left: 15.0,
-      right: 15.0,
-      top: 6.0,
-      bottom: 0.0,
-    );
     final dataLength = NumberFormat.compact(
             locale: Localizations.localeOf(context).toLanguageTag())
         .format(osmData.length);
@@ -44,40 +39,13 @@ class SettingsPage extends ConsumerWidget {
       body: SettingsList(
         contentPadding: EdgeInsets.symmetric(vertical: 10.0),
         sections: [
+          ModeSection(),
           SettingsSection(
-            title: 'Editing Modes',
-            titlePadding: titlePadding,
+            title: Text(loc.settingsApiServer),
             tiles: [
               SettingsTile(
-                title: 'Buildings and Entrances',
-                trailing: Icon(Icons.navigate_next),
-                onPressed: (context) {
-                  Navigator.pushReplacement(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => EntranceEditorPage()),
-                  );
-                },
-              ),
-              SettingsTile.switchTile(
-                title: loc.settingsMicromapping,
-                subtitle: 'Benches, trees, and street lamps',
-                onToggle: (value) {
-                  ref
-                      .read(editorModeProvider.notifier)
-                      .set(value ? EditorMode.micromapping : EditorMode.poi);
-                },
-                switchValue: editorMode == EditorMode.micromapping,
-              ),
-            ],
-          ),
-          SettingsSection(
-            title: loc.settingsApiServer,
-            titlePadding: titlePadding,
-            tiles: [
-              SettingsTile(
-                title: loc.settingsLoginDetails,
-                subtitle: login,
+                title: Text(loc.settingsLoginDetails),
+                description: Text(login ?? ''),
                 trailing: Icon(Icons.navigate_next),
                 onPressed: (context) {
                   Navigator.push(
@@ -89,11 +57,10 @@ class SettingsPage extends ConsumerWidget {
             ],
           ),
           SettingsSection(
-            title: loc.settingsDataManagement,
-            titlePadding: titlePadding,
+            title: Text(loc.settingsDataManagement),
             tiles: [
               SettingsTile(
-                title: loc.settingsPurgeData,
+                title: Text(loc.settingsPurgeData),
                 trailing: osmData.length == 0 ? null : Text(dataLength),
                 enabled: osmData.length > 0,
                 onPressed: (_) async {
@@ -103,7 +70,7 @@ class SettingsPage extends ConsumerWidget {
                 },
               ),
               SettingsTile(
-                title: loc.settingsUploads,
+                title: Text(loc.settingsUploads),
                 enabled: changes.length > 0,
                 trailing: changes.length == 0
                     ? null
@@ -118,11 +85,10 @@ class SettingsPage extends ConsumerWidget {
             ],
           ),
           SettingsSection(
-            title: loc.settingsPresentation,
-            titlePadding: titlePadding,
+            title: Text(loc.settingsPresentation),
             tiles: [
               SettingsTile(
-                title: loc.settingsBackground,
+                title: Text(loc.settingsBackground),
                 onPressed: (context) {
                   Navigator.push(
                     context,
@@ -132,22 +98,43 @@ class SettingsPage extends ConsumerWidget {
                 },
               ),
               SettingsTile.switchTile(
-                title: 'Prefer "contact:" prefix',
+                title: Text('Prefer "contact:" prefix'),
                 onToggle: (value) {
                   ref
                       .read(editorSettingsProvider.notifier)
                       .setPreferContact(value);
                 },
-                switchValue: editorSettings.preferContact,
+                initialValue: editorSettings.preferContact,
               ),
               SettingsTile.switchTile(
-                title: 'Fix Numeric Keyboard',
+                title: Text('Fix Numeric Keyboard'),
                 onToggle: (value) {
                   ref
                       .read(editorSettingsProvider.notifier)
                       .setFixNumKeyboard(value);
                 },
-                switchValue: editorSettings.fixNumKeyboard,
+                initialValue: editorSettings.fixNumKeyboard,
+              ),
+              SettingsTile(
+                title: Text('Default Payment Cards'),
+                value: Text(editorSettings.defaultPayment.join(', ')),
+                trailing: Icon(Icons.navigate_next),
+                onPressed: (context) async {
+                  final locale = Localizations.localeOf(context);
+                  final combo = await ref.read(presetProvider).getField('payment_multi', locale);
+                  final List<String>? newValues = await Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => ComboChooserPage(
+                        combo as ComboPresetField,
+                        editorSettings.defaultPayment,
+                      ),
+                    ),
+                  );
+                  if (newValues != null && newValues.isNotEmpty) {
+                    ref.read(editorSettingsProvider.notifier).setDefaultPayment(newValues);
+                  }
+                },
               ),
             ],
           ),